[REFACTOR] Moved UI into separate file
