+++ app/src/main/java/dev/hossain/remotenotify/ui/alertmediumlist/NotificationMediumListScreen.kt
@@ -15,13 +15,11 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -31,14 +29,11 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Slider
-import androidx.compose.material3.SliderDefaults
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
@@ -359,88 +354,6 @@ private fun NotifierCard(
 }
 
 @Composable
-private fun WorkerConfigCard(
-    state: NotificationMediumListScreen.State,
-    modifier: Modifier = Modifier,
-) {
-    var intervalSliderValue by remember { mutableFloatStateOf(state.workerIntervalMinutes.toFloat()) }
-    Card(
-        modifier = modifier,
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Icon(
-                    // painter = painterResource(id = R.drawable.schedule_24dp),
-                    imageVector = Icons.Default.Refresh,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.secondary,
-                    modifier = Modifier.size(32.dp),
-                )
-                Spacer(modifier = Modifier.width(16.dp))
-                Text(
-                    text = "Check Frequency",
-                    style = MaterialTheme.typography.titleMedium,
-                )
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            Text(
-                text = "Alert checked every ${formatDuration(intervalSliderValue.toInt())}",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.secondary,
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Slider(
-                value = intervalSliderValue,
-                onValueChange = { intervalValue: Float ->
-                    intervalSliderValue = intervalValue // Update UI immediately
-                    state.eventSink(NotificationMediumListScreen.Event.OnWorkerIntervalUpdated(intervalValue.toLong()))
-                },
-                valueRange = 30f..300f,
-                // steps = 270, // (300-30)/1 to have steps of 1 minute
-                modifier = Modifier.fillMaxWidth(),
-                colors =
-                    SliderDefaults.colors(
-                        // Increase contrast for the inactive track
-                        inactiveTrackColor = MaterialTheme.colorScheme.surfaceContainerLow,
-                        // Make active part more prominent
-                        activeTrackColor = MaterialTheme.colorScheme.primary,
-                        thumbColor = MaterialTheme.colorScheme.primary,
-                    ),
-            )
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-            ) {
-                Text(
-                    text = "30m",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                Text(
-                    text = "5h",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-        }
-    }
-}
-
-@Composable
 private fun EmptyMediumState(modifier: Modifier = Modifier) {
     Column(
         modifier = modifier.padding(32.dp),
@@ -460,22 +373,6 @@ private fun EmptyMediumState(modifier: Modifier = Modifier) {
         )
     }
 }
-
-@Composable
-private fun formatDuration(minutes: Int): String =
-    when {
-        minutes < 60 -> "$minutes ${if (minutes == 1) "minute" else "minutes"}"
-        minutes % 60 == 0 -> {
-            val hours = minutes / 60
-            "$hours ${if (hours == 1) "hour" else "hours"}"
-        }
-        else -> {
-            val hours = minutes / 60
-            val remainingMinutes = minutes % 60
-            "$hours ${if (hours == 1) "hour" else "hours"} and " +
-                "$remainingMinutes ${if (remainingMinutes == 1) "minute" else "minutes"}"
-        }
-    }
 
 @DrawableRes
 private fun NotifierType.iconResId(): Int =
